+++ appcontainer/AppStateModel.ts
@@ -7,7 +7,6 @@
 import {AppState, AppSuspendData, HoistModel, XH} from '@xh/hoist/core';
 import {action, makeObservable, observable} from '@xh/hoist/mobx';
 import {Timer} from '@xh/hoist/utils/async';
-import {getClientDeviceInfo} from '@xh/hoist/utils/js';
 import {camelCase, isBoolean, isString, mapKeys} from 'lodash';
 
 /**
@@ -24,8 +23,13 @@ export class AppStateModel extends HoistModel {
     suspendData: AppSuspendData;
     accessDeniedMessage: string = 'Access Denied';
 
+    /**
+     * Timestamp when the app first started loading, prior to even JS download/eval.
+     * Read from timestamp set on window within index.html.
+     */
+    readonly loadStarted: number = window['_xhLoadTimestamp'];
+
     private timings: Record<AppState, number> = {} as Record<AppState, number>;
-    private loadStarted: number = window['_xhLoadTimestamp']; // set in index.html
     private lastStateChangeTime: number = this.loadStarted;
 
     constructor() {
@@ -92,7 +96,7 @@ export class AppStateModel extends HoistModel {
                         appBuild: XH.appBuild,
                         locationHref: window.location.href,
                         timings: mapKeys(timings, (v, k) => camelCase(k)),
-                        ...getClientDeviceInfo()
+                        clientHealth: XH.clientHealthService.getReport()
                     },
                     logData: ['appVersion', 'appBuild'],
                     omit: !XH.appSpec.trackAppLoad