+++ jest.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    moduleNameMapper: {
+        '\\.scss$': 'identity-obj-proxy',
+        XHLogo: '<rootDir>/__mocks__/svg.js'
+    },
+    transformIgnorePatterns: ['node_modules/react-markdown']
+};